requirements.txt

streamlit
transformers
sentence-transformers
scikit-learn
pandas
numpy
matplotlib
seaborn
faiss-cpu
tqdm
torch
huggingface_hub
datasets
langchain
openai
python-dotenv
nltk
spacy
PyMuPDF
pdfminer.six
